converted more code
